pyproject.toml
[project]
name = "twelvelabs-nvidia-vss"
version = "0.1.0"
description = "NVIDIA Metropolis VSS + Twelve Labs integration"
requires-python = ">=3.11"
dependencies = [
    "twelvelabs>=0.3.0",
    "aiohttp>=3.9",
    "python-dotenv>=1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
